feat(server): upgrade Go toolchain and dependencies with DNS type definitions - Upgrade Go version from 1.23 to 1.25.5 across workspace modules - Update golang.org/x/crypto, golang.org/x/net and related dependencies - Fix type conversion issues in DNS controller (userIDUint to uint) - Add comprehensive DNS record type constants (A, AAAA, MX, TXT, CNAME, etc.) - Add DNSSEC key type definitions (ZSK, KSK) - Include new golang.org/x/mod, x/sync, x/tools dependencies

+++ server/src/controllers/dns.go
@@ -661,7 +661,7 @@ func (c *DNSController) GetDNSProviderCredentials(ctx *gin.Context) {
 		return
 	}
 
-	credentials, err := c.dnsService.GetDNSProviderCredentials(ctx.Request.Context(), userIDUint, providerID)
+	credentials, err := c.dnsService.GetDNSProviderCredentials(ctx.Request.Context(), uint(userIDUint), providerID)
 	if err != nil {
 		ctx.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
 		return
@@ -707,7 +707,7 @@ func (c *DNSController) CreateDNSProviderCredential(ctx *gin.Context) {
 	}
 
 	credential := &models.DNSProviderCredential{
-		UserID:      userIDUint,
+		UserID:      uint(userIDUint),
 		ProviderID:  req.ProviderID,
 		Name:        req.Name,
 		Credentials: req.Credentials,
